+++ packages/postgrest-react-query/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @supabase-cache-helpers/postgrest-react-query
 
+## 1.11.2
+
+### Patch Changes
+
+- Updated dependencies [b06405b]
+  - @supabase-cache-helpers/postgrest-core@0.8.3
+
+## 1.11.1
+
+### Patch Changes
+
+- Updated dependencies [bfdf3ac]
+  - @supabase-cache-helpers/postgrest-core@0.8.2
+
+## 1.11.0
+
+### Minor Changes
+
+- 74571ce: fix: retain union type in useQuery result
+
 ## 1.10.1
 
 ### Patch Changes